encode: OutputStreamWriter and CommandWriter
Introduce two new classes to extract some common code.

The OutputStreamWriter is inherited by both the capture manager and the
state tracker and provides common utilities for writing to an output
stream.

The CommandWriter is an encode utility for writing commands to an
outpust stream and it is used by both the capture manager and the state
tracker.